+++ src/components/layouts/AppShell.tsx
@@ -10,7 +10,6 @@ import CommandLauncher from "../CommandLauncher";
 import { GlobalAuthPrompt } from "../GlobalAuthPrompt";
 import { SpellbookDropdown } from "../SpellbookDropdown";
 import UserMenu from "../nostr/user-menu";
-import WalletButton from "../WalletButton";
 import { AppShellContext } from "./AppShellContext";
 
 interface AppShellProps {
@@ -77,10 +76,7 @@ export function AppShell({ children, hideBottomBar = false }: AppShellProps) {
             <SpellbookDropdown />
           </div>
 
-          <div className="flex items-center gap-2">
-            <WalletButton />
-            <UserMenu />
-          </div>
+          <UserMenu />
         </header>
         <section className="flex-1 relative overflow-hidden">
           {children}
